requirements.txt
fastapi
uvicorn[standard]
python-socketio
pydantic
sqlalchemy
aiosqlite
python-dotenv